implementing suggestions
